+++ src/main/java/kr/co/easylogin/easyloginwebserver/product/ProductService.java
@@ -3,7 +3,7 @@ package kr.co.easylogin.easyloginwebserver.product;
 import java.util.List;
 import kr.co.easylogin.easyloginwebserver.common.dto.PageDto;
 import kr.co.easylogin.easyloginwebserver.product.domain.Product;
-import kr.co.easylogin.easyloginwebserver.product.dto.ProductInfoResponse;
+import kr.co.easylogin.easyloginwebserver.product.dto.response.ProductInfoResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
